\name{fitted.standrc}
\alias{fitted.standrc}
\title{standrc fitted values}
\description{
Extract fitted values from a standrc object
}
\usage{
\method{fitted}{standrc}(object, ...)
}
\arguments{
  \item{object}{A standrc object}
  \item{...}{further arguments}
}


\author{
Daniel Gerhard
}

\seealso{
\code{\link{standrm}}
}

\keyword{methods}