\name{fixef.standrc}
\alias{fixef.standrc}

\title{standrc fixed effects
}
\description{rstan summary of the fixed effects
}
\usage{
\method{fixef}{standrc}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{A standrc object}
  \item{...}{optional further arguments}
}
\author{
Daniel Gerhard
}  
\keyword{methods}