\name{print.EDsamp}
\alias{print.EDsamp}
\title{Print EDsamp objects}
\description{
  Shows a summary of posterior samples for ED parameters.
}
\usage{
  \method{print}{EDsamp}(x, ...)
}
\arguments{
  \item{x}{An EDsamp object}
  \item{...}{further arguments}
}


\author{
  Daniel Gerhard
}

\seealso{
  \code{\link{standrm}}
}

\keyword{methods}