\name{ranef.standrc}
\alias{ranef.standrc}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{standrc random effects
}
\description{rstan summary of the random effects
}
\usage{
\method{ranef}{standrc}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{A standrc object}
  \item{...}{optional further arguments}
}
\author{
Daniel Gerhard
}  
\keyword{methods}