\name{VarCorr.standrc}
\alias{VarCorr.standrc}

\title{standrc random effect variances}
\description{
Variance parameters of a standrc object
}
\usage{
\method{VarCorr}{standrc}(x, sigma=1, rdig=3)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{A standrc object}
  \item{sigma}{ignored here}
  \item{rdig}{also not used}
}
\author{
Daniel Gerhard
}
\seealso{
\code{\link{standrm}}
}
\keyword{methods}
