\name{waic}
\alias{waic}
\title{Watanabe-Akaike Information Criterion}
\description{
  Calculates WAIC for standrc objects based on Vehtari and Gelman 2014.
}
\usage{
  waic(object)
}
\arguments{
  \item{object}{A standrc object}
}


\author{
  Daniel Gerhard
}

\seealso{
  \code{\link{standrm}}
}

\references{
Vehtari, A., & Gelman, A. (2014). WAIC and cross-validation in Stan. 
Retrieved from http://www.stat.columbia.edu/~gelman/research/unpublished/
}

\keyword{htest}