\name{print.standrc}
\alias{print.standrc}
\title{Print standrc objects}
\description{
Shows a rstan summary of the fixed effects parameters
}
\usage{
\method{print}{standrc}(x, ...)
}
\arguments{
  \item{x}{A standrc object}
  \item{...}{further arguments}
}


\author{
Daniel Gerhard
}

\seealso{
\code{\link{standrm}}
}

\keyword{methods}